refactored to use more interfaces
