+++ cassandra/src/test/java/com/strategicgains/docussandra/persistence/DatabaseRepositoryTest.java
@@ -15,6 +15,7 @@
  */
 package com.strategicgains.docussandra.persistence;
 
+import com.strategicgains.docussandra.persistence.impl.ITableRepositoryImpl;
 import com.strategicgains.docussandra.domain.Database;
 import com.strategicgains.docussandra.domain.Identifier;
 import com.strategicgains.docussandra.testhelper.Fixtures;
@@ -161,7 +162,7 @@ public class DatabaseRepositoryTest
         IndexRepository indexRepo = new IndexRepository(f.getSession());
         assertFalse(indexRepo.exists(Fixtures.createTestIndexOneField().getId()));
         //check iTable deletion
-        ITableRepository iTableRepo = new ITableRepository(f.getSession());
+        ITableRepository iTableRepo = new ITableRepositoryImpl(f.getSession());
         assertFalse(iTableRepo.iTableExists(Fixtures.createTestIndexOneField()));
         //check document deletion
         DocumentRepository docRepo = new DocumentRepository(f.getSession());
