+++ rest/src/main/java/com/strategicgains/docussandra/config/Configuration.java
@@ -24,7 +24,7 @@ import com.strategicgains.docussandra.persistence.DatabaseRepository;
 import com.strategicgains.docussandra.persistence.DocumentRepository;
 import com.strategicgains.docussandra.persistence.IndexRepository;
 import com.strategicgains.docussandra.persistence.IndexStatusRepository;
-import com.strategicgains.docussandra.persistence.QueryRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.QueryRepositoryImpl;
 import com.strategicgains.docussandra.persistence.TableRepository;
 import com.strategicgains.docussandra.service.DatabaseService;
 import com.strategicgains.docussandra.service.DocumentService;
